refactor(MapTool): Moved dataDefinedColumn methods in QgsMapTool
Si they can be reused in other maptools
